requirements.txt
flask
pytailwindcss
lxml
flask-compress
pillow